% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ts_input.R
\name{create_ts_input}
\alias{create_ts_input}
\title{Creates input for time series}
\usage{
create_ts_input(
  start_time,
  end_time,
  span = "30 min",
  run_nighttime = FALSE,
  par_constant,
  par_variable = NULL,
  dir_name = ".",
  filename = "input"
)
}
\description{
This function uses start and end times, a grain of resolution, and constant
and variable parameter values to create the time series input file for SCOPE
time series function. NOTE: Berkeley time format is YYYYMMDDHHMM but this
function only takes input as "\%Y-\%m-\%d \%H:\%M"
}
\details{
\tabular{ll}{
   \strong{Arguments} \tab \strong{Description} \cr
   start_time \tab POSIXct object in "\%Y-\%m-\%d \%H:\%M" format -- this is the first instance for SCOPE to model \cr
   end_time \tab POSIXct object in "\%Y-\%m-\%d \%H:\%M" format -- this is the last instance for SCOPE to model \cr
   span \tab character string that can be read by seq.POSIXct e.g. "1 hour" or "30 min" --  Should be divisible into the difference between start.time and end.time \cr
   run_nighttime \tab logical determining whether SCOPE should run simulations where Rin >= 0 (i.e., at night) \cr
   par_constant \tab a numeric vector of length C -- SCOPE input parameter values that will be repeated for every time step. C is the number of constant variables. \cr
   par_variable \tab T x V matrix -- SCOPE input parameter values change over some or all time steps. T is number of time steps and V is number of varying variables. \cr
   dir_name \tab character string -- name of directory to save the input file \cr
   filename \tab the base.name of the input file to be saved (omitting the '.csv') \cr
}
}
