% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data_default.R
\name{input_data_default}
\alias{input_data_default}
\title{Modifies input_data_default.csv}
\usage{
input_data_default(
  Cdm,
  Cab,
  Cca,
  Vcmax25,
  Cw,
  Cant,
  Rdparam,
  LAI,
  LAT,
  LON,
  timezn
)
}
\description{
The input_dat_default.csv contains input values for leaf (PROSPECT and biochemistry), canopy, soil, meteorology, aerodynamics, time series and angles. This function helps you modify input values for each trait.
At the moment, the function contains leaf biochemical traits and those that are essential for PROSPECT module to run (e.g. Cdm, Cab, Cca, Cw, Cant), indluding LAI, LAT, LON and timezone
}
