% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data_meteo.R
\name{input_data_meteo}
\alias{input_data_meteo}
\title{Modifies input values for Meteorological parameters within input_data_default.csv}
\usage{
input_data_meteo(
  path = "../SCOPE/input/input_data_default.csv",
  z = 5,
  Rin = 600,
  Ta = 20,
  Rli = 300,
  p = 970,
  ea = 15,
  u = 2,
  Ca = 410,
  Oa = 209
)
}
\arguments{
\item{path}{filepath of the filenames.csv}

\item{z}{measurement height of meteorological data (m)}

\item{Rin}{broadband incoming shortwave radiation (0.4-2.5 um) (W m-2)}

\item{Ta}{air temperature (deg C)}

\item{Rli}{broadband incoming longwave radiation (2.5-50 um) (W m-2)}

\item{p}{air pressure (hPa)}

\item{ea}{atmospheric vapour pressure (hPa)}

\item{u}{wind speed at height z (m s-1)}

\item{Ca}{atmospheric CO2 concentration (ppm)}

\item{Oa}{atmospheric O2 concentration (per mille)}
}
\value{
Default values are given based on default SCOPE values in github
}
\description{
This function lets you modify input values for meteorological parameters inside input_data_default.csv contains.
}
