% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_filenames.R
\name{set_filenames}
\alias{set_filenames}
\title{Modify filenames.csv}
\usage{
set_filenames(simulation_name, dataset_dir, meteo_ec_csv, tts)
}
\description{
This function helps you modify filenames.csv
}
\details{
Arguments
simulation_name    Here you can input a string of character to create a simulation name
dataset_dir        Here you can specify the directory of the dataset for the input files
meteo_ec_csv       Here you can specify the name of the .csv file that contains the meteorological information to be used in time series runs
tts                Here you can specify a value for the solar zenith angle. NOTE: this must be equal to blank if timezone, latitude and longitude are specified in the input.filename
}
