% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_options.R
\name{set_options}
\alias{set_options}
\title{Modify setoptions.csv}
\usage{
set_options(verify, simulate)
}
\description{
This function helps you modify setoptions.csv
Arguments
verify       Verifies the results to test the code for the first time. Uses
binary to switch on (1) and off (0) the mode.
simulate     Defines rules for input reading. Values can be 0 (individual runs),
1 (time series runs), or 2 (look-up table)
...
NOTE: more arguments can be added here based on the actual list of parameters
inside setoptions.csv
}
