% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_paramFile.R
\name{set_paramFile}
\alias{set_paramFile}
\title{Modify set_parameter_filenames.csv}
\usage{
set_paramFile(input)
}
\description{
This function helps you modify set_parameter_filenames.csv input file in
matlab for SCOPE to run. It works in three steps:
\enumerate{
\item It reads the set_parameter_filenames.csv file using read_file().
\item It lets you change input file from the default value of
"input_data_latin_hypercube.csv" to "input_data_default.csv" using the
argument "input"
\item It saves the changes using write.table().
}
}
